+++ src/tab/triggers.js
@@ -1,13 +1,13 @@
 import {detectPageLoad, domainExportChecker, domainLoadChecker} from "./checker/domainChecker";
 import {launchExport} from "./scraper/scraper";
 import {uiEnhancer} from "./uiEnhancer/uiEnhancer";
-import {appendDetectClicksModal} from "./uiEnhancer/phind/modals/appendModals";
+import {appendModalDetectClicks} from "./uiEnhancer/modals/appendModals";
 
 export async function actionExtensionIconClicked() {
   const domainPage = await domainExportChecker();
   if (domainPage === null) return;
   launchExport(domainPage);
-  appendDetectClicksModal();
+  appendModalDetectClicks();
 }
 
 export async function actionPageLoaded() {
